Move some extensions methods to the ThirdParty projects
